code/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
python-frontmatter==1.1.0
markdown==3.7
pygments==2.18.0
pydantic-settings==2.11.0
sqlalchemy==2.0.45
openai==1.30.0
cryptography>=42.0.0
google-generativeai>=0.3.0
anthropic>=0.25.0
